feat: switch instance in profile card
