+++ src/components/ProfileCard/ProfileCard.style.ts
@@ -59,5 +59,12 @@ export const useStyles = makeStyles((theme: Theme) =>
       textOverflow: 'ellipsis',
       wordWrap: 'break-word',
     },
+    instance: {
+      marginTop: '10px',
+    },
+    backdrop: {
+      zIndex: theme.zIndex.drawer + 1 + 100,
+      color: '#fff',
+    },
   }),
 );
